LayoutTests/imported/w3c/web-platform-tests/html/webappapis/animation-frames/idlharness-expected.txt

CONSOLE MESSAGE: line 251: callback not yet supported
idlharness test

This test validates the WebIDL included in the Timing control for script-based animations specification.

partial interface Window {
  long requestAnimationFrame(FrameRequestCallback callback);
  void cancelAnimationFrame(long handle);
};

callback FrameRequestCallback = void (DOMHighResTimeStamp time);

PASS Window interface: operation requestAnimationFrame(FrameRequestCallback) 
PASS Window interface: operation cancelAnimationFrame(long) 
PASS Window must be primary interface of window 
PASS Stringification of window 
PASS Window interface: window must inherit property "requestAnimationFrame" with the proper type (0) 
PASS Window interface: calling requestAnimationFrame(FrameRequestCallback) on window with too few arguments must throw TypeError 
PASS Window interface: window must inherit property "cancelAnimationFrame" with the proper type (1) 
PASS Window interface: calling cancelAnimationFrame(long) on window with too few arguments must throw TypeError 
